app/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==5.0.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
Werkzeug==3.0.4
PyMySQL==1.1.1
email-validator==2.2.0
bcrypt==4.2.0

